Add code to detect and break combinational cycles in the timing graph.
The idea behind the implementation of is to identify Strongly
Connected Components (SCCs) in the timing graph which, by definition,
must contain cycles if the include more than one element. This is done using
Tarjan's algorithm in O(V + E) time.  

Once the SCCs are identified an arbitrary edge in the timing graph is 
disconnected to break the cycle. Since it may be possible for smaller sub-SCCs
to result, this is done iteratively until no SCCs with more than one element
are found.