Graphics updates in: general API, text drawing, font caching and UTF-8 support.
Details:
Text in X11 now drawn in colour again.
Added a cache for fonts, improving performance with lots	of
	pointsize-rotaion combinations, and removing text size restrictions
Added support for UTF-8 in all text.
	Note, that postscript will not display multi-byte chars properly
The background now can be specified as a t_color now.
Added getfontsize().
removed normalizing from get_visible_world().
Fix unused out of bounds access, when a dimension of an FPGA is
	smaller than the corrisponding dimension of a clb.